Initial commit of nvim init.vim scripts
Added main init script which contains plugin config for vim development
with golang. There is an additional minimal init script that I used to
debug deoplete and could come in handy again. There is also a script to
be run to bootstrap vim with the plugins, this just needs to be run once
after manually installing vim plug. If you try to start vim before the
plugins are installed the keys are borked and you wont be able to issue
any commands, including exiting.

Also added spell to keep track of custom spellings.