Remove my attempt at having a preview feature
Vscode has a preview feature, that can be useful in certain cases, such
as when navigating some call chain. VSCode will open the first navigated
type in a preview window and then all subsequent navigation (goto
(type/definition/usage) operations will happen in that preview window.
This stops the process of navigating from cluttering the editor with
loads of intermediate files.

I attempted to do this but this attempt didn't work. Nevertheless I
think this coudl be made to work maybe in a completely different way.