Add silent! to organieImport action
Coc nvim changed the way that the organizeImport action worked to throw
an exception rather than return false when failing, this caused it to
display an error in the command bar when failing and also to block
saving. The "silent!" solves this issue and is the recommended way to
configure this.